put starts_with in a separated file
